+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.12.1] - TBA
-* Add Filter By Method, Status Code and Response Size
+* Add Filter By Host, Duration and Content-Type, Method, Status Code and Response Size
 * Wrap Code Viewer in a scrollable container
 
 ## [1.12.0] - 2026-05-01